fix game loop logic error
Fixed an error in the main loop where the round should have
ended but kept going. This was fixed by changing the standard
while loop to a do while loop so that during two player games
the first player's turn is executed once but isn't forced to
continue if only one player is left standing.